move gnutls stuff into its own header/cpp file
