+++ tests/004-streams.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
+#include <quic.hpp>
+#include <quic/gnutls_crypto.hpp>
 #include <thread>
-
-#include "quic.hpp"
 
 namespace oxen::quic::test
 {
