+++ tests/test_client.cpp
@@ -2,9 +2,9 @@
     Test client binary
 */
 
+#include <quic.hpp>
+#include <quic/gnutls_crypto.hpp>
 #include <thread>
-
-#include "quic.hpp"
 
 using namespace oxen::quic;
